INS-647: mock for messagebus in top testutils
