+++ ledger/ledgertestutils/ledgertestutils.go
@@ -17,8 +17,6 @@
 package ledgertestutils
 
 import (
-	"errors"
-	"fmt"
 	"testing"
 
 	"github.com/insolar/insolar/configuration"
@@ -30,54 +28,8 @@ import (
 
 	"github.com/insolar/insolar/ledger"
 	"github.com/insolar/insolar/ledger/storage/storagetest"
+	"github.com/insolar/insolar/testutils/testmessagebus"
 )
-
-type messageBusMock struct {
-	handlers map[core.MessageType]core.MessageHandler
-}
-
-func newMessageBusMock() *messageBusMock {
-	return &messageBusMock{handlers: map[core.MessageType]core.MessageHandler{}}
-}
-
-func (mb *messageBusMock) Register(p core.MessageType, handler core.MessageHandler) error {
-	_, ok := mb.handlers[p]
-	if ok {
-		return errors.New("handler for this type already exists")
-	}
-
-	mb.handlers[p] = handler
-	return nil
-}
-
-func (mb *messageBusMock) MustRegister(p core.MessageType, handler core.MessageHandler) {
-	err := mb.Register(p, handler)
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (mb *messageBusMock) Start(components core.Components) error {
-	panic("implement me")
-}
-
-func (mb *messageBusMock) Stop() error {
-	panic("implement me")
-}
-
-func (mb *messageBusMock) Send(m core.Message) (core.Reply, error) {
-	t := m.Type()
-	handler, ok := mb.handlers[t]
-	if !ok {
-		return nil, errors.New(fmt.Sprint("no handler for message type:", t.String()))
-	}
-
-	return handler(m)
-}
-
-func (mb *messageBusMock) SendAsync(m core.Message) {
-	panic("implement me")
-}
 
 // TmpLedger crteates ledger on top of temporary database.
 // Returns *ledger.Ledger andh cleanup function.
@@ -96,7 +48,7 @@ func TmpLedger(t testing.TB, lr core.LogicRunner, dir string) (*ledger.Ledger, f
 	assert.NoError(t, err)
 
 	// Init components.
-	mb := newMessageBusMock()
+	mb := testmessagebus.NewTestMessageBus()
 	components := core.Components{MessageBus: mb, LogicRunner: lr}
 
 	// Create ledger.
